Windows form
This code create a GUI so that managing personnel can control students accounts by disabling it is its inactive for the specified time from a specific OU(Organizational unit). It also records account activity and stores the data on a log so that this can be stored in the schools database for future reference.